Add error handling (#249)
* Error handling consistent with other discoverers
* Separated discovery of Azure compute and network resources
* Update voc (compute, network)
