+++ service/discovery/azure/compute_test.go
@@ -50,38 +50,6 @@ func (m mockComputeSender) Do(req *http.Request) (res *http.Response, err error)
 				},
 			},
 		}, 200)
-	} else if req.URL.Path == "/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Network/networkInterfaces" {
-		return createResponse(map[string]interface{}{
-			"value": &[]map[string]interface{}{
-				{
-					"id":       "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Network/networkInterfaces/iface1",
-					"name":     "iface1",
-					"location": "eastus",
-					"properties": map[string]interface{}{
-						"networkSecurityGroup": map[string]interface{}{
-							"id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Network/networkSecurityGroups/nsg1",
-						},
-					},
-				},
-			},
-		}, 200)
-	} else if req.URL.Path == "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Network/networkSecurityGroups/nsg1" {
-		return createResponse(map[string]interface{}{
-			"id":       "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Network/networkSecurityGroups/nsg1",
-			"name":     "nsg1",
-			"location": "eastus",
-		}, 200)
-	} else if req.URL.Path == "/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Network/loadBalancers" {
-		return createResponse(map[string]interface{}{
-			"value": &[]map[string]interface{}{
-				{
-					"id":         "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/res1/providers/Microsoft.Network/loadBalancers/lb1",
-					"name":       "lb1",
-					"location":   "eastus",
-					"properties": map[string]interface{}{},
-				},
-			},
-		}, 200)
 	}
 
 	return m.mockSender.Do(req)
@@ -97,20 +65,10 @@ func TestListCompute(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.NotNil(t, list)
-	assert.Equal(t, 3, len(list))
+	assert.Equal(t, 1, len(list))
 
 	storage, ok := list[0].(*voc.VirtualMachineResource)
 
 	assert.True(t, ok)
 	assert.Equal(t, "vm1", storage.Name)
-
-	iface, ok := list[1].(*voc.NetworkInterfaceResource)
-
-	assert.True(t, ok)
-	assert.Equal(t, "iface1", iface.Name)
-
-	lb, ok := list[2].(*voc.LoadBalancerResource)
-
-	assert.True(t, ok)
-	assert.Equal(t, "lb1", lb.Name)
 }
